Revert "chore: Move demo Webpack config to separate file"
This reverts commit 102f6cf5a2baab8a2edcb027ea371db2624e7d18.
