+++ package.json
@@ -9,11 +9,11 @@
     "clean": "del-cli build/** build .closure-tmp/** .closure-tmp",
     "commitmsg": "validate-commit-msg",
     "dist": "npm run build && npm run build:min",
-    "dev": "npm run clean && cross-env MDC_ENV=development webpack-dev-server --progress --content-base demos --inline --hot --host 0.0.0.0 --config=demos/webpack.config.js",
-    "fix:js": "eslint --fix packages test webpack.config.js demos/webpack.config.js karma.conf.js",
+    "dev": "npm run clean && cross-env MDC_ENV=development webpack-dev-server --progress --content-base demos --inline --hot --host 0.0.0.0",
+    "fix:js": "eslint --fix packages test webpack.config.js karma.conf.js",
     "fix:css": "stylelint --fix packages/**/*.scss",
     "fix": "npm-run-all --parallel fix:*",
-    "lint:js": "eslint packages test scripts webpack.config.js demos/webpack.config.js karma.conf.js",
+    "lint:js": "eslint packages test scripts webpack.config.js karma.conf.js",
     "lint:css": "stylelint packages/**/*.scss",
     "lint": "npm-run-all --parallel lint:*",
     "postinstall": "lerna bootstrap",
